Revert "Core/DBLayer: Correct few more wrong read types. Step1"
This reverts commit a3f931a2b98ada9a86d832c27a3e2b15ca79ced8.
DON NOT COMMIT UNFINISHED STUFF. Added in a branch.

+++ src/server/game/Entities/Player/SocialMgr.h
@@ -49,7 +49,7 @@ enum SocialFlag
 struct FriendInfo
 {
     FriendStatus Status;
-    uint8 Flags;
+    uint32 Flags;
     uint32 Area;
     uint8 Level;
     uint8 Class;
@@ -65,7 +65,7 @@ struct FriendInfo
         Note = "";
     }
 
-    FriendInfo(uint8 flags, const std::string& note)
+    FriendInfo(uint32 flags, const std::string& note)
     {
         Status = FRIEND_STATUS_OFFLINE;
         Flags = flags;
